SILKIT-1409: Network Simulator API (#19)
Introducing the SIL Kit Network Simulator API for custom simulation of CAN, LIN, Ethernet and Flexray networks.
On a simulated networks, all controller communication is redirected to a network simulator participant who performs the detailed simulation (e.g. distributing frames and acknowledges, bus protocol specifics, scheduling) and is responsible for the final message distribution. The scope of the API is to manage the message flow for incoming and outgoing messages, not to perform the simulation logic.

+++ SilKit/IntegrationTests/IntegrationTestUtils.hpp
@@ -137,7 +137,7 @@ struct Pipe
         DWORD actualRead = 0;
         buffer_t buf{};
         buf.resize(size);
-        auto ok = ReadFile(handle, buf.data(), buf.size(), &actualRead, nullptr);
+        auto ok = ReadFile(handle, buf.data(), static_cast<DWORD>(buf.size()), &actualRead, nullptr);
 
         if (!ok)
         {
